requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
pandas
numpy
python-dateutil
oracledb
python-dotenv
google-generativeai
tabulate
bcrypt 
flask-mail 
